Move integration tests to unit tests location #review
